backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
motor>=3.5.1
redis>=5.0.0
celery>=5.4.0
google-generativeai
youtube-transcript-api
spacy
sentence-transformers
faiss-cpu
python-dotenv>=1.0.1
pydantic>=2.7.0
python-multipart
